Use connection pool for MySQL queue
