+++ extra-requirements.txt
@@ -1,2 +1,3 @@
 msgpack>=0.5.6
-PyMySQL>=1.0.2
+PyMySQL>=1.0.0
+DBUtils
