+++ tox.ini
@@ -2,6 +2,7 @@
 
 minversion = 2.0
 skipsdist = True
+recreate = false
 envlist = py27, py34, py35, py36, py37, pep8, cover
 deps = -r{toxinidir}/test-requirements.txt
        -r{toxinidir}/extra-requirements.txt
